feature: Add TeamRecruitment API layer and wire MyRecruitmentViewModel

+++ feature/recruitment/src/main/java/in/koreatech/koin/feature/recruitment/ui/myrecruitment/MyRecruitmentState.kt
@@ -11,6 +11,6 @@ data class MyRecruitmentState(
     val posts: ImmutableList<MyRecruitmentPost> = persistentListOf(),
     val filter: RecruitmentFilterState = RecruitmentFilterState(),
     val showCloseDialog: Boolean = false,
-    val closeTargetPostId: Long? = null,
+    val closeTargetPostId: Int? = null,
     val showFilterSheet: Boolean = false
 )
